fix: fixed after lesson 28.11
