+++ src/test/java/com/rakovets/course/java/core/practice/enum_types/DemoEnum.java
@@ -2,12 +2,10 @@ package com.rakovets.course.java.core.practice.enum_types;
 
 public class DemoEnum {
     public static void main(String[] args) {
-        ShowMeSeasons favoriteSeasons = new ShowMeSeasons(SeasonEnum.SUMMER);
+        User favoriteSeasons = new User(Season.SUMMER);
 
-        favoriteSeasons.print(SeasonEnum.SUMMER);
-
-        System.out.println(favoriteSeasons.favouriteSeason(SeasonEnum.SUMMER));
-        System.out.println(favoriteSeasons.getDescription(SeasonEnum.SUMMER));
+        System.out.println(favoriteSeasons.favouriteSeason(Season.SUMMER));
+        System.out.println(favoriteSeasons.getDescription(Season.SUMMER));
         favoriteSeasons.arraySeason();
     }
 }
